[CHORE] #43 : BoardViewController의 Extension 들을 별도 파일로 빼기
